small update for init scripts

+++ init/deploy_configs.sh
@@ -2,8 +2,10 @@ KARABINER_CONF=$HOME/.config/karabiner/karabiner.json
 
 echo "Processing for $KARABINER_CONF"
 if [ -d "$(dirname $KARABINER_CONF)" ]; then
-    echo "> Taking a backup for $KARABINER_CONF with .bk"
-    mv $KARABINER_CONF $KARABINER_CONF.bk
-    echo "> Coping config"
+    if [ -e "$KARABINER_CONF" ]; then
+      echo "> Taking a backup for $KARABINER_CONF with .bk"
+      mv $KARABINER_CONF $KARABINER_CONF.bk
+    fi
+    echo "> Deploying config"
     cp  $HOME/.dotfiles/conf/karabiner.json $KARABINER_CONF
 fi
